singup

+++ 16-10/service/userService.js
@@ -1,4 +1,7 @@
+const STATUS_CODE = require("../constants/httpResponseCode");
+const customerRepo = require("../repositories/customer.repo");
 const userRepo = require("../repositories/user.repo");
+const HelperApp = require("../util/helper");
 const NodeMailerLib = require("../util/nodemailer");
 
 const login = async (data) => {
@@ -22,16 +25,25 @@ const login = async (data) => {
   } catch (e) {}
 };
 
-const signup = async (data) => {
+const signup = async (data, file) => {
   console.log("sign up RUNNING");
 
-  const { name, email, password, dob } = data;
-  console.log("---------------------1-----");
-  // vẻify
+  const { name, email, password, dob, phoneNumber } = data;
+
+  // 1. verify data
+  if (!name || !email || !password || !phoneNumber) {
+    return {
+      error: true,
+      code: STATUS_CODE.badRequest,
+      message: "Nhập đủ thông tin",
+    };
+  }
+
   const user = await userRepo.getUserByCondition({
     email,
     status: 2,
   });
+
   if (user) {
     return {
       error: true,
@@ -39,37 +51,63 @@ const signup = async (data) => {
       message: "User da ton tai",
     };
   }
-  await userRepo.deleteU({
-    email,
+
+  const oldUser = await userRepo.getUserByCondition({
+    email: email,
+    status: 1,
   });
+  if (oldUser) {
+    await userRepo.deleteU({
+      email,
+    });
+    await customerRepo.deleteU({
+      userId: oldUser._id,
+    });
+  }
+
   const newUser = await userRepo.createUser({
     name,
     email,
-    password,
+    password: await HelperApp.hashPw(password),
     dob,
     status: 1,
   });
-  await NodeMailerLib({
-    to: email,
-    subject: "Xác thực k18",
-    text: `http://localhost:3000/auth/signup-confirm?id=${newUser.id}`,
+  console.log(file);
+  const newCustomer = await customerRepo.createCustomer({
+    avatar: file?.filename,
+    phoneNumber,
+    userId: newUser._id,
   });
+
+  const token = HelperApp.generateJwtToken({
+    id: newUser._id,
+  });
+
+  // await NodeMailerLib({
+  //   to: email,
+  //   subject: "Xác thực k18",
+  //   text: `http://localhost:3000/auth/signup-confirm?token=${token}`,
+  // });
+  return {
+    data: token,
+  };
 };
 
 const signupConfirm = async (data) => {
-  const { id } = data;
+  const { token } = data;
+  const { id } = HelperApp.decodeToken(token);
   console.log(id);
   const user = await userRepo.getUserByCondition({
     status: 1,
-    id:id
-  })
+    id: id,
+  });
   if (user) {
     console.log("user", user);
-    await userRepo.update({ _id: id }, { status: 2 })
+    await userRepo.update({ _id: id }, { status: 2 });
   } else {
     return {
-      data:"That bai"
-    }
+      data: "That bai",
+    };
   }
 };
